# back-end/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CreateCustomer": {
				"properties": {
					"phone": {
						"type": "string"
					},
					"surname": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"phone",
					"surname",
					"password",
					"email",
					"name"
				],
				"type": "object"
			},
			"QueryBook": {
				"properties": {
					"stripeId": {
						"type": "string"
					},
					"genres": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"stripeId",
					"genres",
					"price",
					"description",
					"year",
					"author",
					"name",
					"_id"
				],
				"type": "object"
			},
			"CreateBook": {
				"properties": {
					"stripeId": {
						"type": "string"
					},
					"genres": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"stripeId",
					"genres",
					"price",
					"description",
					"year",
					"author",
					"name"
				],
				"type": "object"
			},
			"UpdateBook": {
				"properties": {
					"genres": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"author": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back-end",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/api/registration": {
			"post": {
				"operationId": "Register",
				"responses": {
					"201": {
						"description": "Created"
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Registration",
				"summary": "Registration",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCustomer"
							}
						}
					}
				}
			}
		},
		"/api/books": {
			"get": {
				"operationId": "GetBooks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/QueryBook"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get books",
				"summary": "Get books",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBook",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueryBook"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Server"
					}
				},
				"description": "Create Book",
				"summary": "Create Book",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBook"
							}
						}
					}
				}
			}
		},
		"/api/book/{id}": {
			"get": {
				"operationId": "GetBook",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueryBook"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Get book",
				"summary": "Get book",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteBook",
				"responses": {
					"200": {
						"description": "Success"
					}
				},
				"description": "Delete book",
				"summary": "Delete book",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateBook",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/QueryBook"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Server"
					}
				},
				"description": "Update Book",
				"summary": "Update Book",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBook"
							}
						}
					}
				}
			}
		},
		"/api/books-list": {
			"post": {
				"operationId": "GetBooksByIds",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/QueryBook"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "In the parameters array you need to specify the book IDs",
				"summary": "Get Books from Customer's List by IDs",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"list": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/wish-list/{id}": {
			"get": {
				"operationId": "GetWishList",
				"responses": {
					"201": {
						"description": "Updated"
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Add Book to Customer's List",
				"summary": "Add Book to Customer's List",
				"tags": [
					"Sample Api"
				],
				"security": [],
				"parameters": [
					{
						"description": "The user's identifier",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}